Fixed Form Issue
Clicking on the labels for the input fields automatically redirected
the cursor to the field, but it didn’t work for all of them. Turns out
its because not all the labels “for” names and their corresponding
input “id” names matched. Fixed the issue and changed the button into a
submit input type.